Edit button is updating database and re-rendering Dom
