+++ src/components/news/News.js
@@ -3,7 +3,6 @@ import { Container } from 'reactstrap'
 import NewsList from './NewsList'
 import NewsModal from './NewsModal'
 import API from './../../modules/API/API'
-import moment from 'moment'
 
 export default class News extends React.Component {
 
@@ -36,23 +35,6 @@ export default class News extends React.Component {
     })
   }
 
-  // OnClick functionality to edit saved articles
-  handleEdit = (id) => {
-    // ADD TIMESTAMP
-    let timeSaved = moment(new Date())
-    // ADD USER ID ON CLICK - Currently hard coded, needs to be userSession.getUser()
-    let userId = 2
-    let articleInfo = {
-      title: this.state.title,
-      summary: this.state.summary,
-      url: this.state.url,
-      userId: userId,
-      timestamp: timeSaved
-    }
-    console.log(articleInfo)
-    this.editArticle(articleInfo, id)
-  }
-
   // OnClick functionality to delete saved articles
   handleDelete = (id) => {
     API.deleteData("news", id)
@@ -70,7 +52,7 @@ export default class News extends React.Component {
       <Container>
         <h1 className="text-center mt-5">News Around the 'Berg!</h1>
         <NewsModal saveArticle={this.saveArticle} />
-        <NewsList news={this.state.news} handleDelete={this.handleDelete} handleEdit={this.handleEdit}/>
+        <NewsList news={this.state.news} handleDelete={this.handleDelete} editArticle={this.editArticle}/>
       </Container>)
   }
 }
